Package updates, fix icon issue
Rolled back adwaita-icon-theme to version 41.0 for now, because on 
version 42.0, some category icons in the apps list do not display 
correctly (see screenshot at 
https://cdn.discordapp.com/attachments/845964267520917545/958747504709603408/unknown.png). 
It's not a massive problem since most icons in MassOS are provided by 
Arc; adwaita-icon-theme only has two purposes in MassOS: a dependency of 
GTK3, and fallback icons for ones which Arc doesn't provide.

Also added libsoup3, since GVFS 1.50.0 depends on it. It's not ideal to 
have two different ABI versions of the same program installed, but it 
will do for now, since some software has been ported to libsoup3, and 
other software has not (yet). libsoup (2) will most likely be removed if 
everything becomes ported to libsoup3 in the future.

The patch for Ghostscript 9.56.0 is different to normal patches - this 
is because it's applied as a reverse-patch (the changes in the patch 
have already been applied to the source tree, and we are undoing them). 
This reverse-patch is needed, otherwise building the GS shared library 
fails at the final link on 9.56.0.