Replace Thunderbird with Claws Mail
Thunderbird takes up about 230MB of space, some users have complained 
about it being "bloatware". Claws Mail is only about 5MB and it is more 
integrated with Xfce.

Flatpak is and will be the recommended way to get Thunderbird since it 
allows it to be updated outside of MassOS.

We can also compile Claws without needing to use precompiled binaries.