Replace JS91 with duktape
Polkit 121 added support for using duktape as the JavaScript backend, 
which is far more minimal and lightweight than JS91. No other packages 
in MassOS depend on JS91, so we can remove it completely, saving ~60MB 
of space.

Ports of other desktop environments like GNOME which have packages 
depending on JS91 can have the JS91 build process moved to them (port 
maintainers: please build it the same way it was originally built in 
MassOS).