Significantly reduce the initramfs size
Update Plymouth and dracut to latest git versions, for simpledrm
support. Use of simpledrm in Plymouth removes the need for initramfs to
include GPU firmware, reducing the approximate size from 207M to 69M.
Ubuntu and Fedora have already adopted simpledrm, citing that most of
the historical bugs with simpledrm have now been fixed.

Also, bump SBAT version of grub from grub,4 to grub,5. The SBAT
documentation indicates that grub,5 is the correct entry for GRUB
builds with the fixes for the February 2025 CVEs applied. The git
version of GRUB we use was published long after that, so it will
include those fixes. Newer versions of the Rufus imaging utility will
also produce an annoying warning if the target ISO's GRUB SBAT entry
contains anything below grub,5. So we want to avoid that.