Switch to MS-signed shim
Somehow this fixes a Windows chainloader issue we were having with
newer versions of GRUB. But it's also a better decision in the long
run. Documentation updates in the MassOS wiki will shortly follow. But
in short, it makes it easier to enroll the MassOS secure boot key.
Users should now be able to enroll the key into EITHER the UEFI db
itself, OR simply enroll into MokList using shim's MokManager. The
latter of which is significantly easier on awkward firmwares which do
not allow direct key management via the firmware interface.

Updated GRUB and Linux kernel following this. Generic package updates
will follow in a future commit, but weren't the focus this time around.

Also various package updates for Xfce.