Pull in peimage module for GRUB
It is a downstream module used in Ubuntu and Debian which aims to fix
issues which can be faced when booting EFI applications using GRUB's
chainloader functionality. Pull it in in the hope that it will fix such
issues which can occur in MassOS if secure boot is enabled.

Shim 16.1 (currently a signed version is only available from Fedora) is
supposed to fix this underlying issue itself. Unfortunately it was
found to introduce its OWN set of new issues when trying to boot other
operating systems through GRUB, such as Windows Boot Manager. So
switching to it is not an option unless either Shim or GRUB get fixed.

Note that the GRUB SBAT data must contain a 'grub.peimage,2' entry when
using this module. And is subject to its own revocations if future
security vulnerabilities are found in this module itself - this is
something we are keeping track of along with general GRUB revocations.

Also, miscellaneous (unrelated) package updates.